Convert ShareMessage to higher-order component
